DEPLOY Incluida configuración de devon
Former-commit-id: 0b76a865ec7625b92d9b5f9d34fd7a44046a68bb

+++ config/tool/deploy/cajamar/RECOVERY-DEPLOY-BATCH-SERVER.sh
@@ -1,7 +1,23 @@
 #!/bin/bash
 
+if [ "$#" -lt 1 ]; then
+    echo ""
+    echo "Uso: " $0 " <entorno>"
+    echo ""
+    echo "   <entorno>: desa, pre, pro01, pro02"
+    echo "" 
+    exit 1
+fi
+
+if [[ ! -f config/$1/devon.properties ]] ; then
+    echo ""
+    echo "ERROR: no existe el fichero config/$1/devon.properties"
+    exit 1
+fi
+
 LOCAL_PATH=`pwd`
 
+cp config/$1/devon.properties /recovery/batch-server/
 unzip zip/batch*.zip
 rm -rf /recovery/batch-server/programas/batch/*
 cp -r batch/* /recovery/batch-server/programas/batch/
